Add files via upload
I wrote this sketch with my Son, Reuben Jacob, for his and his friends freshman project.  This is a house-plant watering system, that can water up to 4 plants.
The components are a Arduino Uno, an LCD Keypad Shield, a DS1307 Real Time Clock, a 4 port relay board, and 4 12v DC pumps to water 4 plants.
The design of the code is based on a Finite state machine (FSM).  The design used the design of the alarm clock sketch 
from https://github.com/pAIgn10/AlarmClock, Copyright (c) 2014 Nick Lamprianidis) as a starting point.
The sketch displays the date and time, sets the date and time, and sets the watering times of 4 plants.
There are 4 modes of operation.  The buttons on the LCD have specific functions in each mode.
 <B>Display Date Time mode: </B> Displays Time, and an indication showing if watering is on or of on the LCD (default mode)
        SELECT: Lets you set the watering times of the 4 plants.
        LEFT: Displays the current alarm time
        RIGHT: Turns the Watering system, on and off.  (If off. the set watering times will not activate the motors)
   <B>SHOW Plant Watering Time mode:</B> Displays the times and days of the week set for the different plants
        UP and DOWN: Cycles through the different watering times.   After 5 (configurable) seconds of inactivity, the clock returns to default mode
   <B>Set Watering Time mode: </B>Sets the weekday and watering times of the plants. Day of week, hours & minutes).  After 5 configurable) seconds of inactivity, the clock returns to default mode.
        In weekday mode, LEFT and RIGHT navigates to the weekday.  UP and DOWN selects/ deselects the day
        In Time mode, UP: Increases hours/minutes.  DOWN: Decreases hours/minutes.
        SELECT: Accepts the weekday/hours/minutes value.
        Goes back to default mode after 5 seconds of inactivity
   <B>Watering Time mode:</B> In this mode, waits for the watering time of each plant to be met.  When met, the plants are watered and the buzzer rings